+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 ### Bugfix
 
+- Image loader: fix blurhash canvas aspect ratio #276 @reebalazs
+
 ### Internal
 
 ## 12.3.3 (2022-09-14)